+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 CHANGELOG
 =========
 
-6.1.0
+6.1.0 (2024-04-16)
 ------------------
 
 * Added `PxpFinancial` and `Trustpay` to the `Processor` enum.